requirements.txt
psutil>=5.9.0
gputil>=1.4.0
py-cpuinfo>=9.0.0
numpy>=1.24.0
jinja2>=3.1.0
matplotlib>=3.7.0
plotly>=5.14.0
weasyprint>=59.0
rich>=13.0.0
click>=8.1.0
